Module Fingerboard

module Acoustic_interval : sig ... end

module Cello : sig ... end

A number of utils applicable to the applications of the library to the cello.

module Cents : sig ... end

module Characterized_interval : sig ... end

module Characterized_scale : sig ... end

A characterized scale is a scale in which the intervals between degrees have all been characterized, that is to say their acoustic natures have been explicited in details. This allows to distinguish between, for example, just and pythagorean major scales.

module Fingerboard_location : sig ... end

A fingerboard position characterized by a particular string.

module Fingerboard_position : sig ... end

A position is a fingerboard location that is not characterized by any string, but rather designated that position that is available on all strings at the same relative distance to the open strings. A position is given a name for convenience when in use within a system.

module Frequency : sig ... end

module Interval : sig ... end

module Located_note : sig ... end

A note played at a given location within a system.

module Natural_ratio : sig ... end

A representant of strictly positive ratio of two natural numbers.

module Note : sig ... end

module Roman_numeral : sig ... end

Manipulating small integers displayed as roman numerals. By convention, roman numerals are used to designate the vibrating strings of the instruments, with I (1) being the higher string, and increasing numbers going down. For example, the lower C string on the 4-strings cello is IV (4).

module Scale : sig ... end

module System : sig ... end

module Zarlinean_combination : sig ... end

When playing a just major scale in double stops of thirds or sixths, each pair of notes is expected to form a just interval. Assuming one note is extracted from a pythagorean system (noted p), the other must then be adjusted accordingly. We note z a note taken from a just system to be tuned with the p note.

val hello_world : Core.Sexp.t

val main : Core.Command.t