Module Fingerboard.Cello

A number of utils applicable to the applications of the library to the cello.

module Fingerboard_position_name : sig ... end

val fifth_system : ?acoustic_interval:Acoustic_interval.t -> unit -> System.t

Create a system with 4 strings tuned in fifth and initialise the open_string position. Default tuning the cello to pythagorean fifths, but this may be overrided.

val fingerboard_position : Fingerboard_position_name.t -> Fingerboard_position.t

val add_fingerboard_position_exn : ?on_n_octaves:int -> System.t -> Fingerboard_position_name.t -> unit

val fingerboard_highest_note : Note.t

The selection of this particular note is somewhat arbitrary, anything around that note is realistic here. This is used to limit the generation of scales and positions.

val find_fingerboard_position_exn : System.t -> Fingerboard_position_name.t -> Fingerboard_position.t