Module Fingerboard.Characterized_interval

type t = private {

interval : Interval.t;

acoustic_interval : Acoustic_interval.t;

}

val sexp_of_t : t -> Sexplib0.Sexp.t

val allowed_deviation_from_equal_tempered_12_in_cents : Base.float

A number representing, in cents, the maximum deviation that is allowed as between an acoustic interval and its 12-equal tempered equivalent. Trying to create a t exceeding this bound will result in create_exn to raise.

val create_exn : interval:Interval.t -> acoustic_interval:Acoustic_interval.t -> t