Module Fingerboard.Fingerboard_position

A position is a fingerboard location that is not characterized by any string, but rather designated that position that is available on all strings at the same relative distance to the open strings. A position is given a name for convenience when in use within a system.

A position may serve for playing several notes (with systems with enharmony or tempered systems). Some system may also have distinct positions to play the same notes, using slightly different pitches depending on which scale or tonality it belongs, so there is no one-to-one match between a note and a position.

type t

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : t Base__Ppx_hash_lib.hash_fold

val hash : t -> Base__Ppx_hash_lib.Std.Hash.hash_value

val sexp_of_t : t -> Sexplib0.Sexp.t

val name : t -> Base.string

Retrieve the name supplied with t was created.

val to_string : t -> Base.string

A string identifier constituted from the name and the octave designation.

val ascii_table_columns : t Ascii_table.Column.t Base.list

val acoustic_interval_to_the_open_string : t -> Acoustic_interval.t

val at_octave : t -> octave:Base.int -> t

Returns the same position some number of octaves higher.

val create_exn : name:Base.string -> acoustic_interval_to_the_open_string:Acoustic_interval.t -> t

Expected to be given a vibrating string portion that is between the open string and the first octave. Other positions at higher octave may be constructed using at_octave.

val open_string : t