Module Fingerboard.Roman_numeral

Manipulating small integers displayed as roman numerals. By convention, roman numerals are used to designate the vibrating strings of the instruments, with I (1) being the higher string, and increasing numbers going down. For example, the lower C string on the 4-strings cello is IV (4).

type t =

| I

| II

| III

| IV

| V

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

include Ppx_enumerate_lib.Enumerable.S with type t := t

val all : t list

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : t Base__Ppx_hash_lib.hash_fold

val hash : t -> Base__Ppx_hash_lib.Std.Hash.hash_value

val sexp_of_t : t -> Sexplib0.Sexp.t

val to_string : t -> string

val of_int_exn : int -> t

Because we don't need to handle instruments with a very large number of strings, only a few roman numerals are representable with this module. An overflow will cause the function to raise.

val one : t

val succ_exn : t -> t

val to_int : t -> int